Relocate BluetoothServiceInfo to helpers.service_info (#75195)
