+++ homeassistant/components/bluetooth_tracker/device_tracker.py
@@ -61,7 +61,7 @@ def is_bluetooth_device(device: Device) -> bool:
 def discover_devices(device_id: int) -> list[tuple[str, str]]:
     """Discover Bluetooth devices."""
     try:
-        result = bluetooth.discover_devices(
+        result = bluetooth.discover_devices(  # type: ignore[attr-defined]
             duration=8,
             lookup_names=True,
             flush_cache=True,
@@ -124,7 +124,7 @@ async def get_tracking_devices(hass: HomeAssistant) -> tuple[set[str], set[str]]
 def lookup_name(mac: str) -> str | None:
     """Lookup a Bluetooth device name."""
     _LOGGER.debug("Scanning %s", mac)
-    return bluetooth.lookup_name(mac, timeout=5)  # type: ignore[no-any-return]
+    return bluetooth.lookup_name(mac, timeout=5)  # type: ignore[attr-defined,no-any-return]
 
 
 async def async_setup_scanner(
@@ -180,7 +180,7 @@ async def async_setup_scanner(
             if tasks:
                 await asyncio.wait(tasks)
 
-        except bluetooth.BluetoothError:
+        except bluetooth.BluetoothError:  # type: ignore[attr-defined]
             _LOGGER.exception("Error looking up Bluetooth device")
 
     async def update_bluetooth(now: datetime | None = None) -> None:
